Now does not raise a warning when the file has already been deleted.

+++ main/modules/collections/import.act.php
@@ -259,7 +259,8 @@ class importAction extends MainWindowAction {
 			}
 			
 		}
-		unlink($newName);
+		if (file_exists($newName))
+			unlink($newName);
 		
 		if ($hasErrors) {	
 			$centerPane->add(new Block(ob_get_contents(), 1));
